# src/main/resources/META-INF/openapi.yaml
openapi: 3.1.0
info:
  title: calendar.social API
  version: 1.0.0
  description: |
    Read and write access to calendar.social using an indefinite API token.
    A token has the live permissions of its owning account, has no scopes, and remains valid until revoked.
servers:
  - url: /
security:
  - bearerToken: []
...
paths:
...
  /api/v1/calendars:
    get:
      tags: [Calendars]
      operationId: listCalendars
      summary: List calendars
      responses:
        "200":
          description: All calendars the token owner can edit.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalendarSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Calendars]
      operationId: createCalendar
      summary: Create a calendar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCalendarInput"
      responses:
        "201":
          description: Calendar created. The token owner is its first admin.
          headers:
            ETag:
              $ref: "#/components/headers/CalendarEntityTag"
            Location:
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Calendar"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationFailed"
  /api/v1/calendars/{calendarId}:
    parameters:
      - $ref: "#/components/parameters/CalendarId"
    get:
      tags: [Calendars]
      operationId: getCalendar
      summary: Get a calendar
      responses:
        "200":
          description: Calendar details.
          headers:
            ETag:
              $ref: "#/components/headers/CalendarEntityTag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Calendar"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Calendars]
      operationId: updateCalendar
      summary: Replace calendar settings
      description: Admin access is required.
      parameters:
        - $ref: "#/components/parameters/CalendarIfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalendarSettingsInput"
      responses:
        "200":
          description: Updated calendar.
          headers:
            ETag:
              $ref: "#/components/headers/CalendarEntityTag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Calendar"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "428":
          $ref: "#/components/responses/PreconditionRequired"
  /api/v1/calendars/{calendarId}/calendar-link-regenerations:
    parameters:
      - $ref: "#/components/parameters/CalendarId"
    post:
      tags: [Calendars]
      operationId: regenerateCalendarLink
      summary: Regenerate the compact calendar link
      description: Admin access is required. The previous link stops working immediately.
      parameters:
        - $ref: "#/components/parameters/CalendarIfMatch"
      responses:
        "200":
          description: Calendar with its new public URL.
          headers:
            ETag:
              $ref: "#/components/headers/CalendarEntityTag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Calendar"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "428":
          $ref: "#/components/responses/PreconditionRequired"
...
  /api/v1/calendars/{calendarId}/editor-invitations:
    parameters:
      - $ref: "#/components/parameters/CalendarId"
    post:
      tags: [Invitations]
      operationId: createEditorInvitation
      summary: Create an editor invitation
      responses:
        "201":
          description: Seven-day, single-use editor invitation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invitation"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
      bearerFormat: Opaque API token
      description: Indefinite, unscoped token issued in account settings and valid until revoked.
  parameters:
    CalendarId:
      name: calendarId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
...
    CalendarIfMatch:
      name: If-Match
      in: header
      required: true
      description: Strong ETag from the latest calendar representation.
      schema:
        type: string
        example: '"3"'
...
  headers:
    CalendarEntityTag:
      description: Strong version tag for calendar writes.
      schema:
        type: string
        example: '"3"'
...
  responses:
    Unauthorized:
      description: Missing, malformed, revoked, or unknown bearer token.
      headers:
        WWW-Authenticate:
          schema:
            type: string
            example: Bearer realm="calendar.social"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Forbidden:
      description: The token owner does not have the required live calendar role.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFound:
      description: The resource does not exist or is deliberately hidden from this account.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    PreconditionFailed:
      description: The resource changed since the supplied ETag was read.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    PreconditionRequired:
      description: A current If-Match header is required.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    ValidationFailed:
      description: The JSON document is structurally valid but contains invalid application data.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    CalendarSummary:
      type: object
      required: [id, name, role, publicAccessEnabled, publicUrl]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        role:
          $ref: "#/components/schemas/CalendarRole"
        publicAccessEnabled:
          type: boolean
        publicUrl:
          type: string
          format: uri
    Calendar:
      allOf:
        - $ref: "#/components/schemas/CalendarSummary"
        - type: object
          required: [description, timeZone, version]
          properties:
            description:
              type: [string, "null"]
              maxLength: 4000
            timeZone:
              type: string
              description: IANA time-zone identifier.
            version:
              type: integer
              minimum: 0
    CreateCalendarInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 160
    CalendarSettingsInput:
      type: object
      required: [name, timeZone, publicAccessEnabled]
      properties:
        name:
          type: string
          maxLength: 160
        description:
          type: [string, "null"]
          maxLength: 4000
        timeZone:
          type: string
          maxLength: 80
        publicAccessEnabled:
          type: boolean
    CalendarRole:
      type: string
      enum: [EDITOR, ADMIN]
...
    Invitation:
      type: object
      required: [id, kind, calendarId, url, createdAt, expiresAt]
      properties:
        id:
          type: integer
          format: int64
        kind:
          type: string
          enum: [registration, calendar-editor]
        calendarId:
          type: [integer, "null"]
          format: int64
        calendarName:
          type: [string, "null"]
        url:
          type: string
          format: uri
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
...
    Problem:
      type: object
      required: [type, title, status, detail]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri-reference